pyproject.toml
[tool.poetry]
name = "cli_tools"
version = "0.1.0"
description = ""
authors = ["Alexander Lewzey <a.lewzey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydub = "^0.25.1"
simpleaudio = "^1.0.4"
soundfile = "^0.12.1"
scipy = "^1.13.0"
matplotlib = "^3.8.4"
tqdm = "^4.66.2"
fire = "^0.6.0"
transformers = "^4.40.1"
torch = "^2.3.0"
moviepy = "^1.0.3"
pillow = "9.5.0"
pydantic = "^2.10.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.3.2"


[tool.poetry.group.test.dependencies]
ipykernel = "^6.29.4"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli_tools = "cli_tools.cli:main"



[tool.ruff]
select = [
    "A", # Pyflakes: general checks
    "B", # pycodestyle: pycodestyle checks
    "C", # mccabe: complexity checks
    # "D",  # pydocstyle: docstring style checks
    "E",  # pycodestyle: pycodestyle errors
    "F",  # pyflakes: unused imports, variables, etc.
    "I",  # isort: import sorting
    "N",  # pep8-naming: naming conventions
    "S",  # bandit: security checks
    "W",  # pycodestyle: pycodestyle warnings
    "UP", # pyupgrade: upgrade python version syntax
]

ignore = [
    "S101", # Ignore use of assert detected
    "S108", # ignore use of tmp
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]
fix = true
